model_deployment/requirements.txt
flask
scikit-learn
db-sqlite3
redis
pandas
requests
gunicorn
python-dotenv